rename file to document poller to line up with what it is actually doing
