+++ src/services/ui/src/features/package-actions/lib/modules/issue-rai/index.tsx
@@ -12,7 +12,7 @@ import {
   AttachmentsSection,
   PackageSection,
 } from "@/components";
-import { CheckDocumentFunction } from "@/utils/Poller/seaStatusPoller";
+import { CheckDocumentFunction } from "@/utils/Poller/documentPoller";
 import { SEATOOL_STATUS } from "shared-types";
 
 export const defaultIssueRaiSchema = z.object({
