+++ src/services/ui/src/features/package-actions/lib/modules/temporary-extension/index.tsx
@@ -14,7 +14,7 @@ import {
   AttachmentsSection,
 } from "@/components";
 import { TEPackageSection } from "@/features/package-actions/lib/modules/temporary-extension/legacy-components";
-import { CheckDocumentFunction } from "@/utils/Poller/seaStatusPoller";
+import { CheckDocumentFunction } from "@/utils/Poller/documentPoller";
 
 export const defaultTempExtSchema = z
   .object({
