+++ src/services/ui/src/features/submission/spa/chip-intitial.tsx
@@ -32,7 +32,7 @@ import {
 
 import { useQuery as useQueryString } from "@/hooks";
 import { SlotAdditionalInfo } from "@/features";
-import { documentPoller } from "@/utils/Poller/seaStatusPoller";
+import { documentPoller } from "@/utils/Poller/documentPoller";
 
 const formSchema = z.object({
   id: zSpaIdSchema,
